+++ src/game/items/Weapon.java
@@ -6,9 +6,9 @@ import game.Handler;
 
 public abstract class Weapon extends Item {
 
-	protected int bps; //How many bullets per second
+	protected double bps; //How many bullets per second
 	
-	public Weapon(BufferedImage texture, String name, int bps, String desc1, String desc2, String desc3, String desc4, String desc5, String desc6, String desc7, int id) {
+	public Weapon(BufferedImage texture, String name, double bps, String desc1, String desc2, String desc3, String desc4, String desc5, String desc6, String desc7, int id) {
 		super(texture, name, desc1, desc2, desc3, desc4, desc5, desc6, desc7, id);
 		this.bps = bps;
 	}
@@ -18,7 +18,7 @@ public abstract class Weapon extends Item {
 
 	
 	//GETTERS & SETTERS
-	public int getBps() {
+	public double getBps() {
 		return bps;
 	}
